Bug 26959: (follow-up) Fix CSS for column selection labels
This patch modifies the staff interface SCSS in order to fix the display
of column selection checkbox labels on the batch item modification page.
Even though the style of the elements defines the "white-space" as
"nowrap," adding whitespace within the element (as was done to correct
indentation) still created problems.

Changing the display property of the elements to "inline-block" corrects
this.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>